+++ libs/react/src/__mocks__/sidebarMock.ts
@@ -1,10 +1,9 @@
 export function getSidebarMock() {
   return {
-    id: '66b2ff76-39cb-4a0d-83c7-951d34f3f653',
+    id: 'cc0267d8-f49c-493e-8ea0-2aaa58bb61f3',
     name: 'sidebar_navigation',
+    type: 'NAVIGATION_SIDE',
     description: 'Provides links in the application sidebar',
-    created_at: '2023-05-21T19:47:11.390Z',
-    updated_at: '2023-05-21T19:47:11.969Z',
     definition: {
       items: [
         {
@@ -16,28 +15,42 @@ export function getSidebarMock() {
           route: '/home',
         },
         {
-          key: 'footprint_key',
-          icon: {
-            name: 'ColdFootprintIcon',
-          },
-          label: 'Footprint',
-          route: '/footprint',
-        },
-        {
-          key: 'journey_key',
-          icon: {
-            name: 'ColdJourneyIcon',
-          },
-          label: 'Journey',
-          route: '/journey',
-        },
-        {
           key: 'compliance_key',
           icon: {
             name: 'ColdComplianceIcon',
           },
           label: 'Compliance',
           route: '/compliance',
+        },
+        {
+          key: 'assessments_key',
+          icon: {
+            name: 'ColdJourneyIcon',
+          },
+          label: 'Assessments',
+          route: '/assessments',
+        },
+        {
+          key: 'actions_key',
+          icon: {
+            name: 'ColdActionsIcon',
+          },
+          label: 'Actions',
+          route: '/actions',
+        },
+        {
+          key: 'reports_key',
+          icon: {
+            name: 'ColdReportIcon',
+          },
+          label: 'Reports',
+          items: [
+            {
+              key: 'reports_carbon_footprint_key',
+              label: 'Carbon Footprint',
+              route: '/reports/carbon_footprint',
+            },
+          ],
         },
         {
           key: 'documents_key',
@@ -48,70 +61,34 @@ export function getSidebarMock() {
           route: '/documents',
         },
         {
-          key: 'actions_key',
-          icon: {
-            name: 'ColdActionsIcon',
-          },
-          items: [
-            {
-              key: 'overview_actions_key',
-              label: 'Overview',
-              route: '/actions',
-            },
-            {
-              key: 'facilities_actions_key',
-              label: 'Facilities',
-              route: '/actions/facilities',
-            },
-            {
-              key: 'travel_actions_key',
-              label: 'Travel',
-              route: '/actions/travel',
-            },
-            {
-              key: 'operations_actions_key',
-              label: 'Operations',
-              route: '/actions/operations',
-            },
-            {
-              key: 'product_actions_key',
-              label: 'Product',
-              route: '/actions/product',
-            },
-            {
-              key: 'employee_footprint_actions_key',
-              label: 'Employee Footprint',
-              route: '/actions/employee_footprint',
-            },
-            {
-              key: 'employee_activation_actions_key',
-              label: 'Employee Activation',
-              route: '/actions/employee_activation',
-            },
-            {
-              key: 'internal_alignment_actions_key',
-              label: 'Internal Alignment',
-              route: '/actions/internal_alignment',
-            },
-            {
-              key: 'community_impact_actions_key',
-              label: 'Community Impact',
-              route: '/actions/community_impact',
-            },
-          ],
-          label: 'Actions',
-        },
-        {
           key: 'settings_key',
-          placement: 'bottom',
           icon: {
             name: 'ColdSettingsIcon',
           },
+          items: [
+            {
+              key: 'settings_company_info_key',
+              label: 'Company Info',
+              route: '/settings/company_info',
+            },
+            {
+              key: 'settings_account_key',
+              label: 'Account',
+              route: '/settings/account',
+            },
+            {
+              key: 'settings_user_key',
+              label: 'Users',
+              route: '/settings/user',
+            },
+          ],
           label: 'Settings',
-          route: '/settings',
           roles: ['cold:admin', 'company:admin', 'company:owner'],
+          route: '/settings',
         },
       ],
     },
+    created_at: '2023-09-11T17:17:02.295Z',
+    updated_at: '2024-02-12T10:49:58.323Z',
   };
 }