+++ libs/react/src/components/application/application/application.tsx
@@ -1,15 +1,25 @@
 import { ColdRoutes } from '../routes';
 import { matchRoutes, useLocation } from 'react-router-dom';
 import { GuidanceButton } from '../../molecules';
+import { useFlags } from 'launchdarkly-react-client-sdk';
 
 export const Application = () => {
   const location = useLocation();
-  const shouldRenderGuidanceButton = matchRoutes([{ path: '/' }, { path: '/home' }, { path: '/footprint' }, { path: '/journey' }, { path: '/actions/*' }], location);
+  const ldFlags = useFlags();
+  const shouldRenderGuidanceButton = () => {
+    let routes = [];
+    if (ldFlags.showReiComplianceMvpSidebarCold506) {
+      routes = [{ path: '/' }, { path: '/home' }, { path: '/compliance/*' }, { path: '/assessments/*' }, { path: '/actions/*' }, { path: '/reports/*' }, { path: '/documents' }];
+    } else {
+      routes = [{ path: '/' }, { path: '/home' }, { path: '/footprint' }, { path: '/journey' }, { path: '/actions/*' }, { path: '/compliance/*' }, { path: '/documents' }];
+    }
+    return matchRoutes(routes, location.pathname);
+  };
 
   return (
     <div className="max-w-[1440px] m-auto overflow-x-clip">
       <ColdRoutes />
-      {shouldRenderGuidanceButton && <GuidanceButton />}
+      {shouldRenderGuidanceButton() && <GuidanceButton />}
     </div>
   );
 };